+++ src/main/java/com/kakaotech/team14backend/inner/point/usecase/GetPointUsecase.java
@@ -3,8 +3,8 @@ package com.kakaotech.team14backend.inner.point.usecase;
 import com.kakaotech.team14backend.common.MessageCode;
 import com.kakaotech.team14backend.exception.MemberNotFoundException;
 import com.kakaotech.team14backend.inner.member.model.Member;
-import com.kakaotech.team14backend.inner.point.model.GetPointPolicy;
-import com.kakaotech.team14backend.inner.point.model.Point;
+import com.kakaotech.team14backend.point.domain.GetPointPolicy;
+import com.kakaotech.team14backend.point.domain.Point;
 import com.kakaotech.team14backend.inner.point.repository.PointRepository;
 import javax.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
