+++ src/main/java/com/kakaotech/team14backend/inner/point/usecase/UsePointUsecase.java
@@ -2,7 +2,7 @@ package com.kakaotech.team14backend.inner.point.usecase;
 
 import com.kakaotech.team14backend.common.MessageCode;
 import com.kakaotech.team14backend.exception.MemberNotFoundException;
-import com.kakaotech.team14backend.inner.point.model.Point;
+import com.kakaotech.team14backend.point.domain.Point;
 import com.kakaotech.team14backend.inner.point.model.PointHistory;
 import com.kakaotech.team14backend.inner.point.model.TransactionType;
 import com.kakaotech.team14backend.inner.point.repository.PointHistoryRepository;
