+++ src/main/java/com/kakaotech/team14backend/point/domain/GetPointPolicy.java
@@ -1,4 +1,4 @@
-package com.kakaotech.team14backend.inner.point.model;
+package com.kakaotech.team14backend.point.domain;
 
 public enum GetPointPolicy {
