+++ src/main/java/com/kakaotech/team14backend/post/application/GetPopularPostPoint.java
@@ -1,7 +1,7 @@
 package com.kakaotech.team14backend.post.application;
 
 import com.kakaotech.team14backend.common.RedisKey;
-import com.kakaotech.team14backend.inner.point.model.UsePointDecider;
+import com.kakaotech.team14backend.point.domain.UsePointDecider;
 import com.kakaotech.team14backend.post.dto.PostLevelPoint;
 import com.kakaotech.team14backend.post.domain.PostLevel;
 import lombok.RequiredArgsConstructor;
