+++ src/test/java/com/kakaotech/team14backend/inner/post/usecase/GetPopularPostPointUsecaseTest.java
@@ -1,6 +1,6 @@
 package com.kakaotech.team14backend.inner.post.usecase;
 
-import com.kakaotech.team14backend.inner.point.model.UsePointDecider;
+import com.kakaotech.team14backend.point.domain.UsePointDecider;
 import com.kakaotech.team14backend.post.dto.PostLevelPoint;
 import com.kakaotech.team14backend.post.application.GetPopularPostPoint;
 import com.kakaotech.team14backend.post.application.SavePopularPosts;
